src/dashboard/requirements.txt
streamlit
google-cloud-firestore
pandas
plotly